fix: address review issues in ScreenTerminalOutputStream refactoring
- Fix ByteBuffer offset bug preserving undecoded bytes across flushes
- Extract DelegateOutputStream to ScreenTerminalOutputStream to eliminate
  duplication between SwingTerminal and WebTerminal
- Use StandardCharsets.UTF_8 instead of Charset.defaultCharset() in Tmux
